Refactor ViewTOS-structure
